Updated module augmentation for font package.
