+++ packages/ckeditor5-font/src/fontbackgroundcolor/fontbackgroundcolorui.ts
@@ -38,9 +38,3 @@ export default class FontBackgroundColorUI extends ColorUI {
 		return 'FontBackgroundColorUI';
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ FontBackgroundColorUI.pluginName ]: FontBackgroundColorUI;
-	}
-}
