+++ packages/ckeditor5-font/src/fontcolor/fontcolorcommand.ts
@@ -29,9 +29,3 @@ export default class FontColorCommand extends FontCommand {
 		super( editor, FONT_COLOR );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		[ FONT_COLOR ]: FontColorCommand;
-	}
-}
