+++ packages/ckeditor5-font/src/fontconfig.ts
@@ -388,41 +388,3 @@ export interface FontSizeOption {
 	 */
 	upcastAlso?: Array<MatcherPattern>;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface EditorConfig {
-
-		/**
-		 * The configuration of the font background color feature.
-		 * It is introduced by the {@link module:font/fontbackgroundcolor/fontbackgroundcolorediting~FontBackgroundColorEditing} feature.
-		 *
-		 * Read more in {@link module:font/fontconfig~FontColorConfig}.
-		 */
-		fontBackgroundColor?: FontColorConfig;
-
-		/**
-		 * The configuration of the font color feature.
-		 * It is introduced by the {@link module:font/fontcolor/fontcolorediting~FontColorEditing} feature.
-		 *
-		 * Read more in {@link module:font/fontconfig~FontColorConfig}.
-		 */
-		fontColor?: FontColorConfig;
-
-		/**
-		 * The configuration of the font family feature.
-		 * It is introduced by the {@link module:font/fontfamily/fontfamilyediting~FontFamilyEditing} feature.
-		 *
-		 * Read more in {@link module:font/fontconfig~FontFamilyConfig}.
-		 */
-		fontFamily?: FontFamilyConfig;
-
-		/**
-		 * The configuration of the font size feature.
-		 * It is introduced by the {@link module:font/fontsize/fontsizeediting~FontSizeEditing} feature.
-		 *
-		 * Read more in {@link module:font/fontconfig~FontSizeConfig}.
-		 */
-		fontSize?: FontSizeConfig;
-	}
-}
-
