+++ packages/ckeditor5-font/src/fontfamily/fontfamilycommand.ts
@@ -29,9 +29,3 @@ export default class FontFamilyCommand extends FontCommand {
 		super( editor, FONT_FAMILY );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		[ FONT_FAMILY ]: FontFamilyCommand;
-	}
-}
