+++ packages/ckeditor5-font/src/fontfamily/fontfamilyediting.ts
@@ -14,8 +14,6 @@ import FontFamilyCommand from './fontfamilycommand';
 import { normalizeOptions } from './utils';
 import { buildDefinition, FONT_FAMILY } from '../utils';
 
-import '../fontconfig';
-
 /**
  * The font family editing feature.
  *
@@ -131,11 +129,5 @@ export default class FontFamilyEditing extends Plugin {
 				value: ( viewElement: ViewElement ) => viewElement.getAttribute( 'face' )
 			}
 		} );
-	}
-}
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ FontFamilyEditing.pluginName ]: FontFamilyEditing;
 	}
 }
