+++ packages/ckeditor5-font/src/fontsize/fontsizeediting.ts
@@ -15,8 +15,6 @@ import FontSizeCommand from './fontsizecommand';
 import { normalizeOptions } from './utils';
 import { buildDefinition, FONT_SIZE, type FontConverterDefinition } from '../utils';
 
-import '../fontconfig';
-
 // Mapping of `<font size="..">` styling to CSS's `font-size` values.
 const styleFontSize = [
 	'x-small', // Size "0" equal to "1".
@@ -190,11 +188,5 @@ export default class FontSizeEditing extends Plugin {
 				}
 			}
 		} );
-	}
-}
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ FontSizeEditing.pluginName ]: FontSizeEditing;
 	}
 }
